Add a pipeline for Character appearances
There's a few missing pieces to this, but it shows a full pipeline that
can be picked up by the changes to the engine and used ingame.

Need to include the actual bit where we use the input image as a
template and replace pixels in it with colours defined in the
character-appearance config.